Changed to_f by to_i

+++ convert_minutes/lib/convert_minutes.rb
@@ -4,7 +4,7 @@ class ConvertMinutes
   attr_accessor :n
 
   def initialize(n = nil)
-    @n = n.to_f
+    @n = n.to_i
   end
 
   def converter
